+++ packages/engine/Source/Scene/ArcGisMapServerImageryProvider.js
@@ -16,9 +16,12 @@ import RuntimeError from "../Core/RuntimeError.js";
 import TileProviderError from "../Core/TileProviderError.js";
 import WebMercatorProjection from "../Core/WebMercatorProjection.js";
 import WebMercatorTilingScheme from "../Core/WebMercatorTilingScheme.js";
+import ArcGisMapService from "./ArcGisMapService.js";
 import DiscardMissingTileImagePolicy from "./DiscardMissingTileImagePolicy.js";
 import ImageryLayerFeatureInfo from "./ImageryLayerFeatureInfo.js";
 import ImageryProvider from "./ImageryProvider.js";
+import ArcGisBaseMapType from "./ArcGisBaseMapType.js";
+import DeveloperError from "../Core/DeveloperError.js";
 
 /**
  * @typedef {object} ArcGisMapServerImageryProvider.ConstructorOptions
@@ -39,8 +42,7 @@ import ImageryProvider from "./ImageryProvider.js";
  *        that no tiles are discarded, construct and pass a {@link NeverTileDiscardPolicy} for this
  *        parameter.
  * @property {boolean} [usePreCachedTilesIfAvailable=true] If true, the server's pre-cached
- *        tiles are used if they are available.  If false, any pre-cached tiles are ignored and the
- *        'export' service is used.
+ *        tiles are used if they are available. Exporting Tiles is only supported with deprecated APIs.
  * @property {string} [layers] A comma-separated list of the layers to show, or undefined if all layers should be shown.
  * @property {boolean} [enablePickFeatures=true] If true, {@link ArcGisMapServerImageryProvider#pickFeatures} will invoke
  *        the Identify service on the MapServer and return the features included in the response.  If false,
@@ -59,6 +61,8 @@ import ImageryProvider from "./ImageryProvider.js";
  * @property {number} [tileHeight=256] The height of each tile in pixels.  This parameter is ignored when accessing a tiled server.
  * @property {number} [maximumLevel] The maximum tile level to request, or undefined if there is no maximum.  This parameter is ignored when accessing
  *                                        a tiled server.
+ *
+ *
  */
 
 /**
@@ -85,6 +89,7 @@ function ImageryProviderBuilder(options) {
     credit = new Credit(credit);
   }
   this.credit = credit;
+  this.tileCredits = undefined;
   this.tileDiscardPolicy = options.tileDiscardPolicy;
 
   this.tileWidth = defaultValue(options.tileWidth, 256);
@@ -104,6 +109,7 @@ ImageryProviderBuilder.prototype.build = function (provider) {
   provider._tilingScheme = this.tilingScheme;
   provider._rectangle = this.rectangle;
   provider._credit = this.credit;
+  provider._tileCredits = this.tileCredits;
   provider._tileDiscardPolicy = this.tileDiscardPolicy;
   provider._tileWidth = this.tileWidth;
   provider._tileHeight = this.tileHeight;
@@ -221,7 +227,11 @@ function metadataSuccess(data, imageryProviderBuilder) {
   }
 
   if (defined(data.copyrightText) && data.copyrightText.length > 0) {
-    imageryProviderBuilder.credit = new Credit(data.copyrightText);
+    if (defined(imageryProviderBuilder.credit)) {
+      imageryProviderBuilder.tileCredits = [new Credit(data.copyrightText)];
+    } else {
+      imageryProviderBuilder.credit = new Credit(data.copyrightText);
+    }
   }
 }
 
@@ -255,7 +265,7 @@ async function requestMetadata(resource, imageryProviderBuilder, provider) {
   });
 
   try {
-    const data = await jsonResource.fetchJsonp();
+    const data = await jsonResource.fetchJson();
     metadataSuccess(data, imageryProviderBuilder);
   } catch (error) {
     metadataFailure(resource, error, provider);
@@ -264,34 +274,49 @@ async function requestMetadata(resource, imageryProviderBuilder, provider) {
 
 /**
  * <div class="notice">
- * To construct a ArcGisMapServerImageryProvider call {@link ArcGisMapServerImageryProvider.fromUrl}. Do not call the constructor directly.
+ * This object is normally not instantiated directly, use {@link ArcGisMapServerImageryProvider.fromBasemapType} or {@link ArcGisMapServerImageryProvider.fromUrl}.
  * </div>
  *
  * Provides tiled imagery hosted by an ArcGIS MapServer.  By default, the server's pre-cached tiles are
  * used, if available.
+ * 
+ * <br/>
+ * 
+ * An {@link https://developers.arcgis.com/documentation/mapping-apis-and-services/security| ArcGIS Access Token } is required to authenticate requests to an ArcGIS Image Tile service.
+ * To access secure ArcGIS resources, it's required to create an ArcGIS developer
+ * account or an ArcGIS online account, then implement an authentication method to obtain an access token.
  *
  * @alias ArcGisMapServerImageryProvider
  * @constructor
  *
  * @param {ArcGisMapServerImageryProvider.ConstructorOptions} [options] Object describing initialization options
  *
- * @see ArcGisMapServerImagery.fromUrl
- * @see BingMapsImageryProvider
- * @see GoogleEarthEnterpriseMapsProvider
- * @see OpenStreetMapImageryProvider
- * @see SingleTileImageryProvider
- * @see TileMapServiceImageryProvider
- * @see WebMapServiceImageryProvider
- * @see WebMapTileServiceImageryProvider
- * @see UrlTemplateImageryProvider
+ * @see ArcGisMapServerImageryProvider.fromBasemapType
+ * @see ArcGisMapServerImageryProvider.fromUrl
  *
  * @example
+ * // Set the default access token for accessing ArcGIS Image Tile service
+ * Cesium.ArcGisMapService.defaultAccessToken = "<ArcGIS Access Token>";
+ * 
+ * // Add a base layer from a default ArcGIS basemap
+ * const viewer = new Cesium.Viewer("cesiumContainer", {
+ *   baseLayer: Cesium.ImageryLayer.fromProviderAsync(
+ *     Cesium.ArcGisMapServerImageryProvider.fromBasemapType(
+ *       Cesium.ArcGisBaseMapType.SATELLITE
+ *     )
+ *   ),
+ * });
+ *
+ * @example
+ * // Create an imagery provider from the url directly
  * const esri = await Cesium.ArcGisMapServerImageryProvider.fromUrl(
- *     "https://services.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer"
- * );
+ *   "https://ibasemaps-api.arcgis.com/arcgis/rest/services/World_Imagery/MapServer", {
+ *     token: "<ArcGIS Access Token>"
+ * });
  *
  * @see {@link https://developers.arcgis.com/rest/|ArcGIS Server REST API}
- * @see {@link http://www.w3.org/TR/cors/|Cross-Origin Resource Sharing}
+ * @see {@link https://developers.arcgis.com/documentation/mapping-apis-and-services/security| ArcGIS Access Token }
+
  */
 function ArcGisMapServerImageryProvider(options) {
   options = defaultValue(options, defaultValue.EMPTY_OBJECT);
@@ -321,7 +346,13 @@ function ArcGisMapServerImageryProvider(options) {
     this._tilingScheme.rectangle
   );
   this._layers = options.layers;
-  this._credit = undefined;
+  this._credit = options.credit;
+  this._tileCredits = undefined;
+
+  let credit = options.credit;
+  if (typeof credit === "string") {
+    credit = new Credit(credit);
+  }
 
   /**
    * Gets or sets a value indicating whether feature picking is enabled.  If true, {@link ArcGisMapServerImageryProvider#pickFeatures} will
@@ -371,6 +402,117 @@ function ArcGisMapServerImageryProvider(options) {
   }
 }
 
+/**
+ * Creates an {@link ImageryProvider} which provides tiled imagery from an ArcGIS base map.
+ * @param {ArcGisBaseMapType} style The style of the ArcGIS base map imagery. Valid options are {@link ArcGisBaseMapType.SATELLITE}, {@link ArcGisBaseMapType.OCEANS}, and {@link ArcGisBaseMapType.HILLSHADE}.
+ * @param {ArcGisMapServerImageryProvider.ConstructorOptions} [options] Object describing initialization options.
+ * @returns {Promise<ArcGisMapServerImageryProvider>} A promise that resolves to the created ArcGisMapServerImageryProvider.
+ *
+ * @example
+ * // Set the default access token for accessing ArcGIS Image Tile service
+ * Cesium.ArcGisMapService.defaultAccessToken = "<ArcGIS Access Token>";
+ *
+ * // Add a base layer from a default ArcGIS basemap
+ * const provider = await Cesium.ArcGisMapServerImageryProvider.fromBasemapType(
+ *   Cesium.ArcGisBaseMapType.SATELLITE);
+ *
+ * @example
+ * // Add a base layer from a default ArcGIS Basemap
+ * const viewer = new Cesium.Viewer("cesiumContainer", {
+ *   baseLayer: Cesium.ImageryLayer.fromProviderAsync(
+ *     Cesium.ArcGisMapServerImageryProvider.fromBasemapType(
+ *       Cesium.ArcGisBaseMapType.HILLSHADE, {
+ *         token: "<ArcGIS Access Token>"
+ *       }
+ *     )
+ *   ),
+ * });
+ */
+
+ArcGisMapServerImageryProvider.fromBasemapType = async function (
+  style,
+  options
+) {
+  //>>includeStart('debug', pragmas.debug);
+  Check.defined("style", style);
+  //>>includeEnd('debug');
+
+  options = defaultValue(options, defaultValue.EMPTY_OBJECT);
+  let accessToken;
+  let server;
+  let warningCredit;
+  switch (style) {
+    case ArcGisBaseMapType.SATELLITE:
+      {
+        accessToken = defaultValue(
+          options.token,
+          ArcGisMapService.defaultAccessToken
+        );
+        server = Resource.createIfNeeded(
+          defaultValue(options.url, ArcGisMapService.defaultWorldImageryServer)
+        );
+        server.appendForwardSlash();
+        const defaultTokenCredit = ArcGisMapService.getDefaultTokenCredit(
+          accessToken
+        );
+        if (defined(defaultTokenCredit)) {
+          warningCredit = Credit.clone(defaultTokenCredit);
+        }
+      }
+      break;
+    case ArcGisBaseMapType.OCEANS:
+      {
+        accessToken = defaultValue(
+          options.token,
+          ArcGisMapService.defaultAccessToken
+        );
+        server = Resource.createIfNeeded(
+          defaultValue(options.url, ArcGisMapService.defaultWorldOceanServer)
+        );
+        server.appendForwardSlash();
+        const defaultTokenCredit = ArcGisMapService.getDefaultTokenCredit(
+          accessToken
+        );
+        if (defined(defaultTokenCredit)) {
+          warningCredit = Credit.clone(defaultTokenCredit);
+        }
+      }
+      break;
+    case ArcGisBaseMapType.HILLSHADE:
+      {
+        accessToken = defaultValue(
+          options.token,
+          ArcGisMapService.defaultAccessToken
+        );
+        server = Resource.createIfNeeded(
+          defaultValue(
+            options.url,
+            ArcGisMapService.defaultWorldHillshadeServer
+          )
+        );
+        server.appendForwardSlash();
+        const defaultTokenCredit = ArcGisMapService.getDefaultTokenCredit(
+          accessToken
+        );
+        if (defined(defaultTokenCredit)) {
+          warningCredit = Credit.clone(defaultTokenCredit);
+        }
+      }
+      break;
+    default:
+      //>>includeStart('debug', pragmas.debug);
+      throw new DeveloperError(`Unsupported basemap type: ${style}`);
+    //>>includeEnd('debug');
+  }
+
+  return ArcGisMapServerImageryProvider.fromUrl(server, {
+    ...options,
+    token: accessToken,
+    credit: warningCredit,
+    usePreCachedTilesIfAvailable: true, // ArcGIS Base Map Service Layers only support Tiled views
+  });
+};
+
 function buildImageResource(imageryProvider, x, y, level, request) {
   let resource;
   if (imageryProvider._useTiles) {
@@ -413,7 +555,6 @@ function buildImageResource(imageryProvider, x, y, level, request) {
       queryParameters: query,
     });
   }
-
   return resource;
 }
 
@@ -941,7 +1082,7 @@ ArcGisMapServerImageryProvider.prototype.getTileCredits = function (
   y,
   level
 ) {
-  return undefined;
+  return this._tileCredits;
 };
 
 /**
@@ -1078,4 +1219,5 @@ ArcGisMapServerImageryProvider.prototype.pickFeatures = function (
     return result;
   });
 };
+ArcGisMapServerImageryProvider._metadataCache = {};
 export default ArcGisMapServerImageryProvider;