+++ packages/engine/Source/Scene/Cesium3DTile.js
@@ -917,8 +917,9 @@ Cesium3DTile.prototype.getScreenSpaceError = function (
     frameState.mode === SceneMode.SCENE2D ||
     frustum instanceof OrthographicFrustum
   ) {
-    if (defined(frustum._offCenterFrustum)) {
-      frustum = frustum._offCenterFrustum;
+    const offCenterFrustum = frustum.offCenterFrustum;
+    if (defined(offCenterFrustum)) {
+      frustum = offCenterFrustum;
     }
     const pixelSize =
       Math.max(frustum.top - frustum.bottom, frustum.right - frustum.left) /
@@ -1796,7 +1797,7 @@ Cesium3DTile.prototype.createBoundingVolume = function (
  */
 Cesium3DTile.prototype.updateTransform = function (parentTransform) {
   parentTransform = defaultValue(parentTransform, Matrix4.IDENTITY);
-  const computedTransform = Matrix4.multiply(
+  const computedTransform = Matrix4.multiplyTransformation(
     parentTransform,
     this.transform,
     scratchTransform